+++ crates/core/src/catalog.rs
@@ -7,10 +7,6 @@ use futures::future::join_all;
 use isahc::{config::RedirectPolicy, prelude::*};
 use serde::Deserialize;
 
-use std::time::Duration;
-
-const DEFAULT_TIMEOUT: u64 = 30;
-
 const CURSE_CATALOG_URL: &str =
     "https://github.com/casperstorm/ajour-catalog/releases/latest/download/curse.json";
 const TUKUI_CATALOG_URL: &str =
@@ -24,7 +20,6 @@ pub async fn get_catalog_addons_from(url: &str) -> Vec<CatalogAddon> {
     let client = HttpClient::builder()
         .redirect_policy(RedirectPolicy::Follow)
         .max_connections_per_host(6)
-        .timeout(Duration::from_secs(DEFAULT_TIMEOUT))
         .build()
         .unwrap();
 
@@ -94,6 +89,7 @@ pub struct Catalog {
 #[serde(rename_all = "camelCase")]
 #[derive(Debug, Clone, Deserialize, Eq, PartialEq, Ord, PartialOrd)]
 pub struct GameVersion {
+    #[serde(with = "null_to_default")]
     pub game_version: String,
     pub flavor: Flavor,
 }
@@ -101,18 +97,39 @@ pub struct GameVersion {
 #[serde(rename_all = "camelCase")]
 #[derive(Debug, Clone, Deserialize)]
 pub struct CatalogAddon {
+    #[serde(with = "null_to_default")]
     pub id: i32,
+    #[serde(with = "null_to_default")]
     pub website_url: String,
     #[serde(with = "date_parser")]
     pub date_released: Option<DateTime<Utc>>,
+    #[serde(with = "null_to_default")]
     pub name: String,
+    #[serde(with = "null_to_default")]
     pub categories: Vec<String>,
+    #[serde(with = "null_to_default")]
     pub summary: String,
+    #[serde(with = "null_to_default")]
     pub number_of_downloads: u64,
     pub source: Source,
+    #[serde(with = "null_to_default")]
     #[deprecated(since = "0.4.4", note = "Please use game_versions instead")]
     pub flavors: Vec<Flavor>,
+    #[serde(with = "null_to_default")]
     pub game_versions: Vec<GameVersion>,
+}
+
+mod null_to_default {
+    use serde::{self, Deserialize, Deserializer};
+
+    pub fn deserialize<'de, D, T>(deserializer: D) -> Result<T, D::Error>
+    where
+        D: Deserializer<'de>,
+        T: Default + Deserialize<'de>,
+    {
+        let opt = Option::deserialize(deserializer)?;
+        Ok(opt.unwrap_or_default())
+    }
 }
 
 mod date_parser {
@@ -176,5 +193,17 @@ mod tests {
                 panic!("{}", e);
             }
         });
+    }
+
+    #[test]
+    fn test_null_fields() {
+        let tests = [
+            r"[]",
+            r#"[{"id": null,"websiteUrl": null,"dateReleased":"2020-11-20T02:29:43.46Z","name": null,"summary": null,"numberOfDownloads": null,"categories": null,"flavors": null,"gameVersions": null,"source":"curse"}]"#,
+        ];
+
+        for test in tests.iter() {
+            serde_json::from_str::<Vec<CatalogAddon>>(test).unwrap();
+        }
     }
 }