+++ CHANGELOG.md
@@ -14,8 +14,24 @@ and `Removed`.
 
 ## [Unreleased]
 
+### Added
+
+- Added an option to Backup via the command line. Flavors and backup folder sources
+  can be specified
+  - Pass `ajour backup --help` to get help using this new command
+- Two new themes; Ferra and One Dark
+
+### Changed
+
+- Sorted themes alphabetically in the picker
+
 ### Fixed
 
+- Fixed bug that caused catalog to fail downloading when `null` values existed
+  in the payload
+- Removed timeout for downloading the catalog. Users with slow internet can now
+  fetch the catalog regardless of how long it will take  
+- Fix bug that would cause Ajour to panic if the catalog failed to download
 - Fixed error handling to show better human readable errors in Ajour gui. Errors
   and underlying causes are still logged.